pyproject.toml
[project]
name = "lumen"
version = "0.1.0"
description = "External-model AI chat over Spiritist author corpora with RAG"
requires-python = ">=3.11"
dependencies = [
    "agno>=1.7",
    "faiss-cpu",
    "fastapi",
    "openai",
    "pdfminer.six",
    "psycopg[binary]>=3.1",
    "pydantic",
    "pydantic-settings",
    "pyjwt[crypto]>=2.8",
    "sqlalchemy>=2.0",
    "tiktoken",
    "uvicorn[standard]",
]

[dependency-groups]
dev = [
    "pre-commit",
    "ruff",
]

# ── Ruff — Python linter + formatter ──────────────────────────────────────────

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["backend/app", "backend/tests", "evals"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "RUF", # Ruff-specific rules
]
ignore = [
    "E501",   # line too long — handled by formatter
    "B008",   # do not perform function calls in default arguments (FastAPI uses this)
    "N806",   # variable in function should be lowercase (ML code uses caps for matrices)
    "RUF001", # ambiguous unicode in string — EN DASH is intentional in Portuguese text
    "RUF002", # ambiguous unicode in docstring — same as above
    "RUF003", # ambiguous unicode in comment — same as above
]
unfixable = ["F401"]  # unused imports: flag but don't auto-remove (may be re-exports)

[tool.ruff.lint.isort]
known-first-party = ["app"]
force-sort-within-sections = true

[tool.ruff.lint.per-file-ignores]
"backend/tests/*" = ["S101"]  # assert is fine in tests
"evals/*" = ["T201"]          # print() is fine in eval scripts

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

# ── Pyright ────────────────────────────────────────────────────────────────────

[tool.pyright]
pythonVersion = "3.11"
venvPath = "backend"
venv = "venv"
include = ["backend/app"]
reportMissingModuleSource = false
